feat: reduce referenced colorscheme
Eliminate most of the referenced color schemes and replace them with a
custom dark theme and github for the light theme.

Eventually I might replace the light theme with a custom one, but for
now github is good enough.

Additionally, remove `lsp-colors` as it's no longer maintained and
upgrade `nvim-colorizer` to the current name and configuration.